feat(hover): Added `hover.nvim` plugin
This plugin modifies the native hover windows by allowing for multiple
sources to occupy the same window.

Previously, if I wanted to view both the LSP and Diagnostic hover
windows, I needed to call two separate functions with two different
keybindings.

Now, with this plugin, I can just type `gh` to see Diagnostics (if they
exist) and LSP info. And I can switch between the two using <C-n> and
<C-p>!

I might need to modiy this configuration in the future if it becomes the
case that I need this to activate only on `'LspAttach'`.