feat(nvim): Configured linting via nvim-lint
I recently learned that LSP diagnostics and linting are two completely
different things, which means that they need to be configured separately
within Neovim.

Fortunately, nvim-lint is very similar to conform, which is essentially
the same plugin but for formatting. nvim-lint is also quite nice because
it pushes linting errors into `vim.diagnostic` so that they can be
viewed as inline errors.