feat(nvim): Remapped `hover.nvim` to `K` instead of `gh`
I initially mapped this plugin to show hover windows when pressing `gh`
instead of `K` because that's the mapping that's used in the VS Code
Vim emulator extension.

However, I recently learned that `K` is the default keymap for showing
hover menus. So, as part of my recent initiative to make my
configuration as compatible with an OOB Neovim instance, I've decided to
commit to this remap.