fix(nvim): Completely removed <C-j> and <C-k> blink mappings
I gave it some more thought, and I've decided to completely remove the
`<C-j>` and `<C-k>` mappings that I used for navigating up and down a
`blink` completion list. I became motivated to do this after I tried
out `<C-n>` and `<C-p>` by accident on Brave while searching something
up and it navigated through the search results. This made me realize
that the `<C-n>` and `<C-p>` keybindings are far more ubiquitous than
the `<C-j>` and `<C-k>` mappings, so I should get used to them.

I also changed the `<C-n>` and `<C-p>` keymaps in `blink` to use
`"fallback"` because the default for the `cmdline` were causing
`<C-n>` and `<C-p>` to not work when the `blink` menu wasn't up.