feat(nvim): Added `grapple` extension to Telescope
Now I can view all my tagged files in a Telescope window instead of the
default floating window.

Since `grapple` fulfills my initial desire for the buffer picker, I gave
that keymap to `grapple`'s picker and gave the buffer picker a more
generic "search" keymap.

However, I've quickly come to realize that the `grapple` picker is
*insanely* limited. It seems like the plugin maintainer made it hastily
and didn't have the chance to come back to it.

Unfortunately, I think `grapple` as a whole is now unmaintained, so if I
want to continue using it I might have to create my own `telescope`
picker for it. That could be a really fun project down the road, so I
think I'll give it a shot.