feat(nvim)!: Removed `substitue.exchange` keymap
I've come to learn that the keymap I used for `substitute.exchange` is
also used for opening a link under the cursor. I initially thought this
would be okay, but as I've continued to use Neovim I've found myself
wanting to open links more and more.

I've never really used `substitute.exchange`, so I'm okay with removing
it entirely. I'll leave it commented out for now just so I can remind
myself that it exists. I might want to add it back at some point in the
future.