feat(nvim): Added keymaps and modified LSP settings
First, I added keymaps for copying lines of code up and down, similarly
to how you can in VS Code. This implementation doesn't work as well as
VS Code does, but I don't anticipate using it very much in Neovim
because Vim motions are more reliable for that kind of operation.

Second, I modified one of the settings for `cssmodules_ls`. Now, it can
recommend both camelCase and kebab-case CSS classes in my autocompletion
list. Before, it was converting all of my kebab-case classes to
camelCase, which obviously wasn't useful and didn't work.