feat(nvim): Added keymap for deleting word in Insert mode
While Neovim provides a way of doing this out of the box, I decided to
create an additional keymap that I'm more accustomed to in most writing
programs on Mac. I believe I can perform it by default outside of Tmux,
but this mapping ensures it works no matter what

It's a small QoL keymap, and I believe it's a good example of a place
where I can comfortably stray from defaults.