feat(nvim): added keymap for opening `fugitive`'s summary window
This keymap is supposed to mimic my `<leader>w` keymap, which I use for
saving a file (`<leader>` is used in place of `:`).

I'm not sure how I feel about not lumping this one in with my other `Git`
keymaps, but I feel like it's a bit easier to use than potentially
`<leader>gf` or `<leader>gg`. But we'll see over time.