feat(nvim): added `oil` plugin
I added this plugin because I wanted to experiment with a file explorer
that works just like a normal Neovim buffer. I've heard many wonderful
things about `Oil`, so I'm going to attempt using this first.

An alternative that I might use in the future is `Flyer.nvim`, which is
similar to `Oil` but implements directory trees.

If I find that I don't like this type of file explorer I'll just switch
back to `Telescope`'s file browser extension.