refactor(nvim): asynchronous formatting now only happens with `eslint_d`
This change has been a long time coming.

Since the `eslint_d` formatter can take awhile to cold start and has
occasionally died randomly in the middle of coding sessions, it only
makes sense to use it asynchronously via `conform`'s `format_after_save`
option.

However, this option isn't really snappy when it comes to other
formatters. In fact, sometimes it's even problematic. For instance, when
editing Go files over `sshfs`, `format_after_save` would cause
`goimports` to run twice, which would sometimes cause the first format
to be cancelled part of the way through and mess up the way the code
looks (it's also worth mentioning that I setup Go formatting in this
commit!).

So, in this commit, I've made it so that `format_after_save` is only
used when `eslint_d` is one of the configured formatters for the current
buffer. For any other language/formatter combo, `format_on_save` is used.

It's possible that I may have to adjust the `uses_eslint` function in
the event that other formatters have a similar requirement, but that
will be an easy refactor.