feat(nvim): added new keymaps for formatting with `conform`
I added keymaps for toggling formatting and revised the `<leader>f`
keymap to return to normal mode if you performed ranged formatting in
visual mode.

However, with the latter keymap, it seems like the ranged formatting
feature is broken, at least for `stylua`. I suppose this would also
explain why formatting with the `gq` operator is also janky. I'll
probably submit a GitHub issue for this in due time.