refactor(nvim): choosing file with `Oil` now clears search highlights
One annoying issue I've been facing with `Oil` is that I tend to use `/`
search to find a file I want to navigate to. This works great, however
when I enter that file's buffer, the highlighting persists. This can be
really annoying when a file's name is present throughout a file, since
the search highlight will clutter that file.

To ease this minor painpoint, I ammended the `<CR>` keymap to clear all
search highlights once I enter a file's buffer via `Oil`. This setup
makes the most sense to me considering I only ever use `/` search on a
buffer-to-buffer basis.