refactor(nvim): placed all autocmds in a separate file
Additionally, this commit introduces a new util file where I expose
methods that might be helpful throughout my config.

This is my first time creating a Lua module, so it was fun to write a
function and then use it elsewhere, similar to how I would use the
functions exposed by the plugins that I use.