feat(nvim): add stylua and tailwindcss to lualine LSP ignore list
As a rule of thumb, I've decided to place the following types of LSPs
into Lualine's ignore list, that way they never show on the status bar
and take up valuable real estate:

- Supplementary CSS LSPs (e.g., `css_variables`, `tailwindcss`)
- Formatters that are reported as LSPs (e.g., `stylua`)
- Supplementary LSPs that someone wouldn't consider to be the main LSP
  (e.g., `eslint` in a TypeScript file)

In the future, I might create some sort of menu that shows when you
click on the LSP list that displays all active LSPs, or provide some
sort of indicator that there are multiple ones that are hidden. Or a
toggle!