fix(nvim,tmux): preserve cursor style when exiting Neovim in Tmux
On Fedora Workstation 42, I was encountering an issue where entering and
exiting Neovim while inside Tmux would cause the cursor style to change
to a block, even though through Ghostty I expected it to be a vertical
bar the entire time. This was strange because I wasn't experiencing the
same issue on my Mac.

Fortunately, I found a solution thanks to a person on Reddit who was
having the same issue. The solution involves creating an autocommand
that outputs a terminal escape sequence that converts the cursor to its
default state. I also had to add an extra terminal override to Tmux to
ensure it could read those escape sequences properly.