feat(nvim): add `nvim-lspconfig` as an individual plugin
Now that I understand what `nvim-lspconfig` is doing under the hood
(i.e., applying default LSP configurations), it makes sense to include
it as a standalone, dedicated plugin.

Currently, it only exists as a dependency for other plugins, so I
wouldn't want for those plugins to remove the dependency and suddenly my
LSPs no longer work the way they used to.