feat(nvim): remove `typescript-tools`
I've used this plugin for awhile now, and I never really saw much of a
difference between it and `ts_ls`. Although `typescript-tools` tries
to be faster and more extensible, `ts_ls` essentially does the same
stuff at relatively the same speed.

I'm also removing this ahead of the `tsgo` LSP release. If I remove my
soft dependency on `typescript-tools` now, it'll be even easier to
migrate to `tsgo` in the future.