refactor(nvim.lsp): make auto-fix errors autocmd fail gracefully
This commit overwrites many of the changes introduced in the previous
commit.

Rather than relying on a buffer-local variable to determine whether
Diffview is open, I make a protected call (`pcall`) on
`vim.cmd("LspEslintFixAll")` so that I can intercept any errors that
occur and print out my own succinct error message.

While it isn't optimal for have this command fail in the first place,
this is a better catch-all solution than relying on a buffer-local
variable. With the latter solution, I would've had to add more variables
for other plugins that have similar logic to Diffview, which would've
increased the complexity of this already slightly complex autocmd.