feat(nvim): add LSP and formatting support for Svelte
I don't know how to use Svelte, but I intend to learn sometime soon, so
I figured I would set this up in the meantime.

The only part that I'm not too confident about is the plugin integration
with `ts_ls`. I don't know how to test it to see if it's working, so in
the meantime the code exists as a launch point.