feat(nvim): set diagnostic window max width
This commit also sets the `linebreak` setting to `true`, which causes
all wrapped lines to wrap words rather than individual characters.

This needed to be set globally because long TypeScript errors would end
up wrapping within the diagnostic window, but only the letters would be
wrapped instead of the words. This made it really difficult to read the
error.

Also, using this setting in combination with `breakindent` is basically
necessary because sometimes errors have multiple indentation levels that
need to be preserved between wraps. Case in point, again, TypeScript
errors.