feat(nvim): re-enable Snacks.picker as primary picker
Since Folke is now pushing numerous fixes/features to Snacks and since
Telescope appears to be nearing deprecation, I've decided to replace
Telescope in favor of Snacks.Picker!

Funny enough, this is how I set up my Neovim config earlier this year,
but Snacks.picker was still in it's infancy at the time and not ready to
be daily driven. Fortunately, this is no longer the case.