refactor(nvim): change keymap and layout for Colorschemes Picker
The new keymap is slightly more difficult to use, which makes sense
considering changing the current colorscheme is something you'd rarely
want to do.

This also changes the layout to one that's less obtrusive, however the
"vscode" preset completely breaks the live-preview feature of this
picker. I need to submit a bug ticket for this.