fix(nvim): deprioritize `blink` suggestions from 'emmet_language_server'
When working with JSX/TSX files, it's annoying to see Emmet suggestions
in places that aren't relevant (e.g., anywhere outside the `return`
statement of a component). Unfortunately, neither the LSP nor Neovim
provides a way to configure this behavior.

So, as a compromise, this commit adds a sorting function to `blink.cmp`
that always deprioritizes Emmet suggestions. This will make it so that
inside the "HTML" regions of a JSX/TSX file, Emmet suggestions will appear
first because (usually) nothing from `ts_ls` is suggested here. However,
outside these regions, Emmet suggestions will be pushed to the bottom of
the suggestion list because `ts_ls` suggestions exist.