feat(nvim)!: change `Grapple` toggle tag keymap to `<leader>a`
I'm changing this because I want to install a plugin soon that makes
heavy use of the `<leader>m` keymaps, which is what this command was
originally mapped to.

While this new keymap is inspired by Harpoon's keymap for pinning a
buffer, I think it's also pretty ergonomic and elegant when paired with
the toggle tags keymap, `<leader>;`. Both of these keymaps use the pinky
finger when you use touch typing.