refactor(nvim): removed `dial` quote augends
Although this feature is neat, I had to remove the quote augends for
`dial.nvim` because they often took precedence over the augends I
actually wanted to manipulate. For example, in a string containing the
word "second", the quotes would have changed rather than the word
because of the quote augend.

Fortunately, `surround` allows me to change quotes relatively easily, so
I'll just continue using that for now.