perf(nvim)!: remove `emmet_language_server`
The `emment_language_server` drastically impacts completion and general
editor speed over time as you work on a typescriptreact file.

With `ts_ls` already being as slow as it is, it's not worth keeping this
LSP around if it's going to make things even slower. Maybe when `tsgo`
is released I'll re-enable `emmet` (hence why I've only commented out
the code instead of deleting it).