refactor(nvim): remove unnecessary format function for diagnostics
When I originally set up diagnostics for my configuration, I mainly
copied the way kickstart.nvim did it, which included a `format` function
for `virtual_text`. At the time, I was extremely new to Neovim, so I
didn't think much of the function, assuming it was necessary.

However, after coming back to it recently, I realized that that function
actually *doesn't do anything*. If you read it closely, all it really
does is create a useless table that returns the same result it would've
if no `format` function was present in the first place.

It seems like I need to deeply reevaluate most of my kickstart.nvim code
moving forward, because this was extremely ridiculous.