refactor(nvim): remove ESLint auto-fix on save command
Once again, I've changed my stance on formatting TypeScript/JavaScript
files.

Using the `LspEslintFixAll` autocmd on save is nearly as slow as
`eslint_d`, and it makes working with TypeScript files feel lethargic
compared to other languages.

So, I've come to the conclusion that using the built-in LSP formatter is
the best option when Prettier isn't available. It's faster than
`LspEslintFixAll`, and it works in the same way as Prettier does (i.e.,
via `conform.nvim`).