feat(nvim): cond enable treesitter plugins based on treesitter-cli
Add a utility function that can be used to conditionally enable all
treesitter plugins based on the existence of `tree-sitter-cli` on the
user's computer.

If `tree-sitter-cli` is not installed, an error message is displayed
that informs the user to install it before `treesitter` itself can be
enabled.

This logic allows `treesitter` and all its dependent plugins to fail
gracefully in the event that `tree-sitter-cli` isn't installed.

Without this executable, large error messages will appear incessantly.
In other cases, specific parsers won't install but no error messages will be
thrown.