refactor(nvim)!: remove `treesitter-textobjects` ;/, repeat keymaps
This commit makes it so that treesitter node traversal keymaps (e.g.,
`[m`) are not longer repeatable with `;` and `,`.

I've removed this because not all available traversal keymaps are setup
by the textobjects plugin. For example, traversing Git hunks via `[h` is
provided by `gitsigns`. But since it's not controlled by `textobjects`,
it can't be repeated.

This creates frustrating mental overhead where I must remember which
traversal keymaps are setup by `textobjects`, which ones are native, and
which ones are created by plugins in order to determine whether they're
repeatable. Rather than dealing with that, I'm removing the option to
make them repeatable entirely.