feat(nvim): add `include_surrounding_whitespace` option to textobjects
Setting this option to `true` makes it so that custom treesitter
textobjects work more closely to native textobjects.

For example, with this option set, performing `daa` deletes the outer
attribute and the preceeding whitespace. This is similar to `daw`, which
deletes the outer word and also the preceeding whitespace.

Previously, `daa` didn't delete any whitespace, which made it awkward to
use, as I'm accustomed to the way Vim handles textobjects by default.