refactor(nvim)!: swap auto-format toggle keymaps; use `FormatToggle`
Swap the keymaps used for toggling auto-formatting globally and for the
current buffer.

To me, it makes more sense for the elevated, more consequential version of
this keymap to use a capital "F" rather than a lowercase one.

Additionally, since I toggle auto-formatting on a buffer-to-buffer basis
more often than I do globally, I'd rather have that keymap be easier to
invoke with the lowercase "f".