feat(nvim): show file name in `winbar` upon opening splits
This commit makes it so that the current buffer's filename and modified
status are displayed in the `winbar` location whenever one or more
splits are opened.

Whenever a split is focused, the winbar will be colored in blue,
appearing like a small flag above the current file. The inactive split
will dim the filename and modified status.

When winbars are present, the filename that usually rests within the
statusline will be hidden in favor of the winbars.

Certain buffers won't show the winbar even if splits are active because
they don't have a filename (e.g., "[No Name]" files). Some of these
include `Oil` and `Fugitive` buffers. As I discover more, I will add
them to `winbar_disabled_filetypes`.