fix(nvim): ensure `lualine` filename component is shown when necessary
Before this fix, some buffers that I explicitly set to *not* have a
winbar when splits exist wouldn't show their filename in the statusline.
This was problematic because there should always be a way to see a
buffer's filename, even if it's not displayed via the winbar.

To fix this, I made use of the `disabled_filetypes` option in my
`lualine` config to ensure my custom winbar was hidden for those
filetypes.

This has the added benefit of not overriding another
plugin's winbar implementation. This was happening with `nvim-dap-ui`'s
REPL winbar and my previous implementation for disabling specific
filetypes.