fix(nvim): prevent `Diffview` winbars from being overwritten
Previously, winbars set by `Diffview` were getting overwritten by my
custom winbars, making it difficult to understand what the different
splits in a 3-way merge were representing.

I'm not entirely satisfied with this solution because my custom winbars come
back if I switch to a different tmux session. However, I'm unable to
find a clean solution that resolves this, so I'm going to leave it as is
for now.

In the future, it would probably be desirable to dynamically display
winbars based on whether a plugin supplies its own. Otherwise, I
will probably need to implement further bandaid fixes like this one.