feat(nvim): disable plugin maps via `treesitter-textobjects`
According to `treesitter-textobject`'s updated documentation, this
setting needs to be set to prevent potential keymap race conditions and
conflicts.

Before setting this, I would have some issues with my function call and
method textobjects. Sometimes, the incorrect function or the wrong part
of a function would be selected.

It seems like this new setting fixes this issue, but I'll be sure to
remove it if it causes any issues going forward.