fix(nvim): class and method textobjects are now properly selected
Previously, through `treesitter-textobjects`, I made it so that both the
method and class textobjects would have a linewise selection mode. I was
under the impression that this would only apply to Visual mode, however
it actually applied to every mode.

This caused many issues. For example, whenever I tried to delete an
anonymous function inside a JSX attribute, it would delete the entire
line that the attribute was defined on instead of just the function.
This was happening because the Visual line would select the entirety of
all the lines the function was defined on.

After removing these selection modes, the textobjects work as expected.