feat(nvim): add `vim-matchup` plugin
This plugin makes it so that more textobjects, like HTML tags and
quotes, can be jumped to back and forth with the % sign.

It also adds more motions and keymaps to make Neovim's default
`matchparen` functionality even more useful.

I've made some adjustments to the plugin so that it's more performant
and works more similarly to what Zed provides out of the box.